refactor(rewrite): clarify combined workflow module name
